tests/wpt/web-platform-tests/performance-timeline/tentative/include-frames-from-child-cross-origin-grandchild.sub.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
</body>
<script>
promise_test(() => {
    return new Promise(resolve => {
        performance.clearResourceTimings()

        // Create child iframe with an embedded frame that is same-origin with its parent but cross-origin with the current frame.
        const crossOriginChildFrame = document.createElement('iframe')
        crossOriginChildFrame.src = "http://{{hosts[][www]}}:{{ports[http][0]}}/performance-timeline/resources/parent-frame-with-same-origin-child.html"
        document.body.appendChild(crossOriginChildFrame)

        // Listen for postMessage() from child frame.
        window.addEventListener("message", e => {
            // 0 entries for parent, 4 for child, 2 for grandchild.
            assert_equals(e.data, 6)

            const entries = performance.getEntries(true)

            // 3 entries for parent, 0 for child, 0 for grandchild.
            assert_equals(entries.length, 3)

            resolve()
        })
    })
}, "GetEntries of a Cross-Origin child frame with one Cross-Origin grandchild frame")
</script>
